+++ fabric/src/main/java/net/notcoded/wayfix/fabric/WayFixFabric.java
@@ -1,0 +1,11 @@
+package net.notcoded.wayfix.fabric;
+
+import net.fabricmc.api.ClientModInitializer;
+import net.notcoded.wayfix.common.WayFix;
+
+public class WayFixFabric implements ClientModInitializer {
+    @Override
+    public void onInitializeClient() {
+        WayFix.init();
+    }
+}