+++ fabric/src/main/java/net/notcoded/wayfix/fabric/config/ModMenuIntegration.java
@@ -1,0 +1,13 @@
+package net.notcoded.wayfix.fabric.config;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+import net.notcoded.wayfix.common.config.ModConfig;
+import me.shedaniel.autoconfig.AutoConfig;
+
+public class ModMenuIntegration implements ModMenuApi {
+    @Override
+    public ConfigScreenFactory<?> getModConfigScreenFactory() {
+        return parent -> AutoConfig.getConfigScreen(ModConfig.class, parent).get();
+    }
+}