+++ src/main/java/net/notcoded/wayfix/common/WayFix.java
@@ -1,0 +1,36 @@
+package net.notcoded.wayfix.common;
+
+import me.shedaniel.autoconfig.AutoConfig;
+import me.shedaniel.autoconfig.serializer.GsonConfigSerializer;
+import net.notcoded.wayfix.common.config.ModConfig;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.lwjgl.glfw.GLFW;
+
+public class WayFix {
+    public static final Logger LOGGER = LogManager.getLogger(WayFix.class);
+    public static ModConfig config;
+
+    public static void init() {
+        AutoConfig.register(ModConfig.class, GsonConfigSerializer::new);
+        WayFix.config = AutoConfig.getConfigHolder(ModConfig.class).getConfig();
+    }
+
+    public static boolean isWayland() {
+        try {
+            return GLFW.glfwGetPlatform() == GLFW.GLFW_PLATFORM_WAYLAND;
+        } catch (NoSuchMethodError ignored) { // <3.3.0
+            return false;
+        }
+    }
+
+    public static boolean supportsWayland() {
+        try {
+            return GLFW.glfwPlatformSupported(GLFW.GLFW_PLATFORM_WAYLAND);
+        } catch (NoSuchMethodError ignored) { // <3.3.0
+            LOGGER.warn("WayFix is disabling itself due to the LWJGL Version being too low.");
+            LOGGER.warn("Please update to a LWJGL version such as '3.3.1' or higher.");
+            return false;
+        }
+    }
+}