+++ CHANGES.md
@@ -1,2 +1,3 @@
 changes:
 - fix auto gui scaling looking broken with weird window sizes
+- **add support for forge & neoforge!!!!**